Added add data and auth

+++ internal/entity/user.go
@@ -1,7 +1,7 @@
 package entity
 
 type User struct {
-	ID       uint `gorm:"primarykey"`
+	ID       uint64 `gorm:"primarykey"`
 	Login    string
 	Password string
 }
